[refactor] app: file structure update
