+++ app/src/main/java/uk/ryanwong/catnews/di/NiceDateFormatterModule.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.catnews.app.di
+package uk.ryanwong.catnews.di
 
 import dagger.Module
 import dagger.Provides
