+++ app/src/main/java/uk/ryanwong/catnews/newslist/ui/viewmodel/NewsListViewModel.kt
@@ -13,8 +13,8 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import uk.ryanwong.catnews.R
-import uk.ryanwong.catnews.app.di.DispatcherModule
 import uk.ryanwong.catnews.app.util.ErrorMessage
+import uk.ryanwong.catnews.di.DispatcherModule
 import uk.ryanwong.catnews.newslist.data.repository.NewsListRepository
 import uk.ryanwong.catnews.newslist.ui.NewsListUIState
 import java.util.UUID
