+++ app/src/main/java/uk/ryanwong/catnews/storydetail/data/local/entity/StoryEntityMapper.kt
@@ -2,11 +2,9 @@
  * Copyright (c) 2023. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.catnews.uk.ryanwong.catnews.storydetail.data.local.entity
+package uk.ryanwong.catnews.storydetail.data.local.entity
 
 import uk.ryanwong.catnews.domain.model.storydetail.Story
-import uk.ryanwong.catnews.storydetail.data.local.entity.ContentEntity
-import uk.ryanwong.catnews.storydetail.data.local.entity.StoryEntity
 
 fun StoryEntity.toDomainModel(contentEntities: List<ContentEntity>): Story {
     return Story(
