+++ app/src/test/java/uk/ryanwong/catnews/storydetail/data/local/entity/ContentEntityMapperTest.kt
@@ -4,7 +4,6 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldContainInOrder
 import io.kotest.matchers.shouldBe
 import uk.ryanwong.catnews.domain.model.storydetail.Content
-import uk.ryanwong.catnews.uk.ryanwong.catnews.storydetail.data.local.entity.toDomainModel
 
 internal class ContentEntityMapperTest : FreeSpec() {
